Reorganise Identity components
