+++ identity/webapp/views/components/MyAccount/ChangeEmail.tsx
@@ -14,12 +14,15 @@ import {
   useUpdateUser,
 } from '@weco/identity/hooks/useUpdateUser';
 import { validEmailPattern } from '@weco/identity/utils/validation-patterns';
+import { ChangeDetailsModalContentProps } from '@weco/identity/views/components/ChangeDetailsModal';
 import Loading from '@weco/identity/views/components/Loading';
 import { PasswordInput } from '@weco/identity/views/components/PasswordInput';
+import { StatusAlert } from '@weco/identity/views/components/styled/alert';
 import { FieldMargin } from '@weco/identity/views/components/styled/forms';
-
-import { ChangeDetailsModalContentProps } from './ChangeDetailsModal';
-import { ModalContainer, ModalTitle, StatusAlert } from './MyAccount.styles';
+import {
+  ModalContainer,
+  ModalTitle,
+} from '@weco/identity/views/components/styled/modal';
 
 type ChangeEmailInputs = {
   email: string;
