+++ identity/webapp/views/components/MyAccount/index.tsx
@@ -1,17 +1,4 @@
-import ChangeDetailsModal, {
-  ChangeDetailsModalContentProps,
-} from './ChangeDetailsModal';
 import ChangeEmail from './ChangeEmail';
 import ChangePassword from './ChangePassword';
-import DeleteAccount from './DeleteAccount';
-import UnverifiedEmail from './UnverifiedEmail';
 
-export {
-  ChangeDetailsModal,
-  ChangeEmail,
-  ChangePassword,
-  DeleteAccount,
-  UnverifiedEmail,
-};
-
-export type { ChangeDetailsModalContentProps };
+export { ChangeEmail, ChangePassword };
