--- identity/webapp/views/components/PasswordInput/index.ts
@@ -1,2 +1,0 @@
-export { PasswordInput } from './PasswordInput';
-export { PasswordRules } from './PasswordRules';
