--- identity/webapp/views/components/Registration/index.ts
@@ -1,4 +1,0 @@
-import AccountCreated from './AccountCreated';
-import RegistrationInformation from './RegistrationInformation';
-
-export { AccountCreated, RegistrationInformation };
